Refactor to repository pattern
